fixing some issues while running 'npm run build'
