+++ src/app/(routes)/admin/GuardAdmin.tsx
@@ -1,0 +1,24 @@
+'use client';
+import React, { useContext } from 'react'
+import { redirect } from 'next/navigation';
+import { StoreContext } from '../../../store/StoreProvider';
+
+interface Props{
+  children: React.ReactNode;
+}
+
+const GuardAdmin = ({children}:Props) => {
+  const context:any = useContext(StoreContext);
+  const {authAdmin} = context;
+  if(!authAdmin){
+    redirect('/');
+  }
+
+  return (
+    <>
+      {children}
+    </>
+  )
+}
+
+export default GuardAdmin
