+++ src/app/(routes)/admin/layout.tsx
@@ -1,23 +1,18 @@
-'use client';
-import React, { useContext } from 'react'
-import { redirect } from 'next/navigation';
-import { StoreContext } from '../../../store/StoreProvider';
+"use client";
+import React from 'react';
+import GuardAdmin from './GuardAdmin';
 
 interface Props{
   children: React.ReactNode;
 }
 
 const layout = ({children}:Props) => {
-  const context:any = useContext(StoreContext);
-  const {authAdmin} = context;
-  if(!authAdmin){
-    redirect('/');
-  }
-
   return (
-    <>
+    <GuardAdmin>
+      <>
       {children}
-    </>
+      </>
+    </GuardAdmin>
   )
 }
