+++ src/app/(routes)/home/layout.tsx
@@ -1,22 +1,16 @@
 "use client";
-import React, { useContext } from "react";
+import React from "react";
 import HomePage from "./HomePage";
 import { PrincipalLayout } from "../../principal-layout/PrincipalLayout";
-import { redirect } from "next/navigation";
-import { StoreContext } from "../../../store/StoreProvider";
+import GuardHome from "./GuardHome";
 
 const layout = ({ children }: { children: React.ReactNode }) => {
-  const context:any = useContext(StoreContext);
-  const {authNormal} = context;
-  if(!authNormal){
-    redirect('/');
-  }
   return (
-    <>
+    <GuardHome>
       <PrincipalLayout>
         <HomePage>{children}</HomePage>
       </PrincipalLayout>
-    </>
+    </GuardHome>
   );
 };
